fix: useIPCEvents hook rerendering
The hook useIPCEvents has been refactored to simplify its usage in ContentSettings component. Instead of calling the hook in ContentSettings, it is now called in App and the required method (writeMailContentToTxt) is passed down as a prop to ContentSettings. This change will improve code readability and data flow within the app.