fix: CSS class order in Navigation component
The class order for the list items in the Navigation component has been changed. The 'rounded-full' style has been moved to the front, to ensure that it applies regardless of the route. This way, the components will maintain a uniform shape feature no matter what the state of the application is.