feat: custom scrollbar styles and MailList page
Introduced custom scrollbar styles in main.css and created a new MailList page. Also, made a fix in CSS utility classes for console textarea component to prevent resizing. The Navigation component was updated to include a link to the new MailList page. Note that array elements now join with a semicolon for the arrayToConsoleString function within the array-to-console-string.js file.