fix: last Row generation on deletion
Updated the logic in Table.jsx for managing the mail list rows. Added conditions to check if all rows have been deleted and add an empty one if true, or if the last row is not empty, another empty one is appended. This gives better user experience while editing the mail list.