feat: detected changes indicator
This commit refactor "Textarea" and "MailTitleInput" components. The "useEffect" function in the Textarea component was removed, and a new variable "isChanging" was added to track if there are changes to save in both components. A visual cue was also given to the user when such changes occurred. Finally, the button label in the Textarea component was updated from "Save Content" to "Save Changes".