feat: color-theme support in tailwind
Implemented a new feature for users to switch themes. Updated the application to use the theme selected by the user and store it in local storage. Improved the use of the CSS by moving main.css out of the assets folder, adding themes.css, and adjusting the font URL paths in main.css accordingly.