refactor: color theme and update related component styles
This commit refactors the color theme and updates the styles of related components to align with the new theme. The change includes renaming the color variables, reflecting the new color variables in the components, and updating the Tailwind CSS configuration to use the new color variables. Also, some minor changes have been made to improve the color usage in the components for a better user interface.