chore: package to use lottie-react and remove deprecated dependencies
This commit updates the mailbot-electron package to use the current "lottie-react" library instead of the obsolete "react-lottie". It also removes the deprecated dependencies "babel-runtime" and "core-js" and updates the project version to 1.1.2 in package-lock.json. The changes have also been reflected in the 'Button.jsx' file where the lottie animation in use has been updated.